Peer cache for query too
