+++ cc-tools-demo.json
@@ -1,7 +1,7 @@
 {
-    "name": "Fail proof",
+    "name": "GoShare-",
     "description": {
-        "ptbr": "Fail proof",
-        "en": "Fail proof"
+        "ptbr": "GoShare-",
+        "en": "GoShare-"
     }
 }
