+++ chaincode/assetTypeList.go
@@ -1,13 +1,12 @@
 package main
 
 import (
-	"github.com/goledgerdev/cc-tools-demo/chaincode/assettypes"
 	"github.com/goledgerdev/cc-tools/assets"
 )
 
 var assetTypeList = []assets.AssetType{
-	assettypes.Person,
-	assettypes.Book,
-	assettypes.Library,
-	assettypes.Secret,
+	//	assettypes.Person,
+	//	assettypes.Book,
+	//	assettypes.Library,
+	//	assettypes.Secret,
 }
