+++ chaincode/txList.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	txdefs "github.com/goledgerdev/cc-tools-demo/chaincode/txdefs"
-
 	tx "github.com/goledgerdev/cc-tools/transactions"
 )
 
@@ -10,7 +8,7 @@ var txList = []tx.Transaction{
 	tx.CreateAsset,
 	tx.UpdateAsset,
 	tx.DeleteAsset,
-	txdefs.CreateNewLibrary,
-	txdefs.GetNumberOfBooksFromLibrary,
-	txdefs.UpdateBookTenant,
+	//	txdefs.CreateNewLibrary,
+	//	txdefs.GetNumberOfBooksFromLibrary,
+	//	txdefs.UpdateBookTenant,
 }
