+++ rest-server/src/fabric-artifacts/sdkCommon/index.ts
@@ -1,4 +1,5 @@
 export { default as getAllPeersSDK } from './getAllPeersSDK';
 export { default as initializeChannel } from './initializeChannel';
+export { default as initializeChannelPeer } from './initializeChannelPeer';
 export { default as createChannelWithOrderer } from './createChannelWithOrderer';
 export { default as createPeerTargets } from './createPeerTargets';
